aimbot-external/requirements.txt
opencv-python
mss
numpy
pydirectinput
pywin32
PyYAML